requirements.txt
# requirements.txt

flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
PyJWT==2.10.1
requests==2.31.0
pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
opennsfw2==0.14.0